move form submission to its own controller
